refactor: change file naming and components placing
