+++ src/components/ErrorButton.tsx
@@ -1,13 +1,13 @@
 import React, { ReactNode } from 'react';
 
-export default function Search_button({
+export default function ErrorButton({
   onClick,
 }: {
   onClick: () => void;
 }): ReactNode {
   return (
-    <button onClick={onClick} className="search_button">
-      SEARCH
+    <button onClick={onClick} className="error_button">
+      THROW AN ERROR
     </button>
   );
 }
