+++ src/components/ErrorMessager.tsx
@@ -1,0 +1,6 @@
+import {ReactNode} from "react";
+
+export default function ErrorMessager (): ReactNode {
+  return (<div className='error'> OOPS! Something went wrong. Please, reload the page </div>);
+}
+
