+++ src/types/Result.ts
